+++ src/api/types/modules.ts
@@ -17,25 +17,15 @@ export interface Feature {
 }
 
 export interface FeatureQuery {
-    feature_id: string;
+    feature_id?: string;
+    query?: string;
+    type?: string;
     source?: string;
     species?: string;
 }
 
-export interface FeaturesQuery {
-    feature_id: string;
-    source?: string;
-    species?: string;
-}
-
-export interface FeatureSearchQuery extends Query {
-    query: string;
-    source?: string;
-    species?: string;
-}
-
-export interface FeatureAutocompleteQuery extends FeatureSearchQuery {
-    query: string;
+export interface FeaturePasteQuery extends FeatureQuery {
+    pasted: string[];
 }
 
 // ------------------------------------------------------------------