+++ docs/CHANGELOG.rst
@@ -5,6 +5,10 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+==========
+Unreleased
+==========
+
 ===================
 15.1.0 - 2021-03-16
 ===================