+++ docs/CHANGELOG.rst
@@ -5,6 +5,12 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+==========
+Unreleased
+==========
+
+- Make FeatureQuery and FeaturesQuery source, species parameters optional.
+
 ===================
 16.0.0 - 2021-11-16
 ===================